feat: disable move if local

+++ build_run.sh
@@ -2,7 +2,10 @@
 
 bun install
 bun run build
-sudo mkdir -p /var/www/qb-poc-preview
-sudo rm -rf /var/www/qb-poc-preview/####VAR:FOLDER_ID####
-sudo mv dist /var/www/qb-poc-preview/####VAR:FOLDER_ID####
-sudo chown -R www-data:www-data /var/www/qb-poc-preview/
+# Check if --local flag is passed
+if [[ "$1" != "--local" ]]; then
+    sudo mkdir -p /var/www/qb-poc-preview
+    sudo rm -rf /var/www/qb-poc-preview/####VAR:FOLDER_ID####
+    sudo mv dist /var/www/qb-poc-preview/####VAR:FOLDER_ID####
+    sudo chown -R www-data:www-data /var/www/qb-poc-preview/
+fi
